fix: poll interval to request Notion
In order to know when the export is finished, the poll interval to Notion
was set to 50ms. Notion didn't bother this one (I guess they didn't pay
attention ever before) but from little time ago, on this very frequent
request a 429 is being received. So, as the export use to take from
seconds to minutes, was tested than a value of 5000ms is ok and
don't bother Notion (for now...).