fix: poll interval to request Notion (#9)
* fix: poll interval to request Notion

In order to know when the export is finished, the poll interval to Notion
was set to 50ms. Notion didn't bother this one (I guess they didn't pay
attention ever before) but from little time ago, on this very frequent
request a 429 is being received. So, as the export use to take from
seconds to minutes, was tested than a value of 5000ms is ok and
don't bother Notion (for now...).

* feat: poll interval in config

Add a property (optional) in the config to set the poll interval
to check if the export has finished (default to 2000ms).

* chore: poll interval in config

Removed as parameter function and make it mandatory in config.

* fix: missing poll interval value

Set poll internal as an optional value in config and added private
constant with a default interval value in the case the config one
is not provided. Be aware it is used _only_ in a private method

* Fix linting and use default config.

---------

Co-authored-by: Guille <guillermo.c.martinez@dcsl.com>
Co-authored-by: Yann Bolliger <yann@bolliger.io>